+++ tests/unit/Infrastructure/Manage/Dto/fixture/manage_entity_saml20_no_excluded_from_push.json
